Add fuzzing test to flowgger
